pyproject.toml
[project]
name = "hotel-matching"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "imagehash>=4.3.2",
    "openai-clip>=1.0.1",
    "pillow>=11.3.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.1",
    "flask>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.32.0",
    "opencv-python-headless>=4.10.0.84",
    "numpy>=2.1.0",
]
